requirements.txt
astroid==1.5.3
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
Django==1.11.7
djangorestframework==3.7.3
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.2
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
virtualenv==15.1.0
wrapt==1.10.11
